Continued working on the  "semi brute force" approach to best graph finding

+++ src/quantumlunch/bestgraph/semibruteforce/ValueSource.java
@@ -1,5 +1,5 @@
 package quantumlunch.bestgraph.semibruteforce;
 
 public interface ValueSource {
-    Long value();
+    Integer value();
 }
